requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
cobble==0.1.4
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.1

fonttools==4.54.1
future==1.0.0

grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httplib2==0.22.0

mysql-connector-python==9.0.0

pillow==10.4.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyMySQL==1.1.1
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
rich==13.8.1
rsa==4.9
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
starlette==0.38.6
thinc==8.2.5
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6

